refactor: extract PreviewPanel from Preview for reuse
